1st draft of register/deregister.  no tests or test fixes yet.
